+++ src/constants.ts
@@ -1,7 +1,10 @@
-import { ClassPathScope } from './enum';
+import { ClassMatchScope } from './enum';
 import { AutoClassPluginSettings } from './interfaces';
 
 export const DEFAULT_SETTINGS: AutoClassPluginSettings = {
-  paths: [{ path: 'Example Path/Subfolder/', classes: ['example-class'], scope: ClassPathScope.Preview }],
-  version: '1.5.3'
+  matches: [
+    { path: 'Example Path/Subfolder/', classes: ['example-class'], scope: ClassMatchScope.Preview },
+    { tag: '#ExampleTag', classes: ['another-class'], scope: ClassMatchScope.Edit }
+  ],
+  version: '2.0.0'
 };