+++ src/enum.ts
@@ -1,4 +1,4 @@
-export enum ClassPathScope {
+export enum ClassMatchScope {
   Preview = 'Preview',
   Edit = 'Edit',
   Both = 'Preview & Edit'